Moved Formatters -> SearchHelper, removed method_missing fallback
