+++ frontend/app/controllers/search_controller.rb
@@ -6,68 +6,6 @@ class SearchController < ApplicationController
   # as JSON or HTML fragments to be rendered into tables in the application,
   # either directly in templates, or as responses to ajax calls.
 
-
-  module Formatter
-    # Formatters are procs that take a fieldname, and return a proc usable as the
-    # value_block argument in SearchHelper::ExtraColumn#new
-    #
-    # Usage:
-    #       Formatter[formatter, field].call(field)
-    #
-    # FORMATTERS is customized to return a default formatter if passed a format it doesn't understand,
-    # which will produce the value as returned from the field without alteration.
-    #
-    # Because these formatters are referenced as methods, names must be valid ruby identifiers
-    class << self
-      def [](meth, field)
-        self.send(meth.to_sym, field)
-      end
-
-      def stringify(field)
-        proc {|record| record[field].to_s }
-      end
-
-      def linked_records_listing(field)
-        proc {|record|
-          identifiers = Array(record[field])
-          out_html = %Q|<ul class="linked-records-listing count-#{identifiers.length}">|
-          out_html << identifiers.map {|identifier|
-            %Q|<li><span class="collection-identifier">#{identifier}</span></li>|
-          }.join("")
-          out_html << '</ul>'
-
-          out_html.html_safe
-        }
-      end
-
-      def combined_identifier(field)
-        proc {|record|
-          identifiers = Array(record['collection_identifier_stored_u_sstr'])
-          displays = Array(record['collection_display_string_u_sstr'])
-          out_html = %Q|<ul class="linked-records-listing count-#{identifiers.length}">|
-          out_html << identifiers.zip(displays).map {|identifier, display|
-            %Q|<li><span class="collection-identifier">#{identifier}</span> <span class="collection-display-string">#{display}</span></li>|
-          }.join('')
-          out_html << '</ul>'
-
-          out_html.html_safe
-        }
-      end
-
-      # Default to stringify, alias method so successive lookups skip method_missing
-      def method_missing(meth, *args)
-        new_name = meth.to_sym
-
-        Rails.logger.warn("Defining default SearchController::Formatter #{new_name}, please define explicit formatter")
-        define_singleton_method(new_name) do |field|
-          stringify(field)
-        end
-        method(new_name).call(args.first)
-      end
-    end
-  end
-
-
   set_access_control  "view_repository" => [:do_search, :advanced_search]
 
   include ExportHelper
@@ -142,7 +80,7 @@ class SearchController < ApplicationController
       @search_data = Search.all(session[:repo_id], params_for_backend_search.merge({"facet[]" => SearchResultData.BASE_FACETS.concat(params[:facets]||[]).uniq}))
       if params[:extra_columns]
         @extra_columns = params[:extra_columns].map do |opts|
-          SearchHelper::ExtraColumn.new(I18n.t(opts['title'], default: opts['title']), Formatter[opts['formatter'], opts['field']], opts['sort_options'] || {}, @search_data)
+          SearchHelper::ExtraColumn.new(I18n.t(opts['title'], default: opts['title']), SearchHelper::Formatter[opts['formatter'], opts['field']], opts['sort_options'] || {}, @search_data)
         end
       end
       @display_identifier = params.fetch(:display_identifier, false) == 'true'
