+++ frontend/app/helpers/search_helper.rb
@@ -286,4 +286,54 @@ module SearchHelper
     end
 
   end
+
+  module Formatter
+    # Formatters are procs that take a fieldname, and return a proc usable as the
+    # value_block argument in SearchHelper::ExtraColumn#new
+    #
+    # Usage:
+    #       Formatter[formatter, field].call(field)
+    #
+    # FORMATTERS is customized to return a default formatter if passed a format it doesn't understand,
+    # which will produce the value as returned from the field without alteration.
+    #
+    # Because these formatters are referenced as methods, names must be valid ruby identifiers
+    class << self
+      def [](meth, field)
+        self.send(meth.to_sym, field)
+      end
+
+      def stringify(field)
+        proc {|record| record[field].to_s }
+      end
+
+      def linked_records_listing(field)
+        proc {|record|
+          identifiers = Array(record[field])
+          out_html = %Q|<ul class="linked-records-listing count-#{identifiers.length}">|
+          out_html << identifiers.map {|identifier|
+            %Q|<li><span class="collection-identifier">#{identifier}</span></li>|
+          }.join("")
+          out_html << '</ul>'
+
+          out_html.html_safe
+        }
+      end
+
+      def combined_identifier(field)
+        proc {|record|
+          identifiers = Array(record['collection_identifier_stored_u_sstr'])
+          displays = Array(record['collection_display_string_u_sstr'])
+          out_html = %Q|<ul class="linked-records-listing count-#{identifiers.length}">|
+          out_html << identifiers.zip(displays).map {|identifier, display|
+            %Q|<li><span class="collection-identifier">#{identifier}</span> <span class="collection-display-string">#{display}</span></li>|
+          }.join('')
+          out_html << '</ul>'
+
+          out_html.html_safe
+        }
+      end
+    end
+  end
+
 end
